+++ src/main/java/com/uah/ismael/portal_formula1/model/entity/Piloto.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class Piloto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String nombre;
@@ -32,12 +32,11 @@ public class Piloto {
     @JoinColumn(name = "equipoID")
     private Equipo equipo;
 
-    // Getters y Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
