+++ src/main/java/com/uah/ismael/portal_formula1/model/entity/Simulacion.java
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 public class Simulacion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String tipo;
@@ -26,12 +26,11 @@ public class Simulacion {
     @JoinColumn(name = "responsableID")
     private Usuario responsable;
 
-    // Getters y Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
